requirements.txt
# LEMM Requirements for HuggingFace Space
# Python 3.10+ required for ACE-Step compatibility
# See .python-version and runtime.txt for Python version specification

# ZeroGPU Support (for H200/A100 spaces)
spaces

# Core AI/ML - Specific versions for reproducibility
torch==2.5.1
torchaudio==2.5.1
torchvision==0.20.1

# Transformers ecosystem - Versions compatible with ACE-Step
# ACE-Step requires transformers==4.50.0, diffusers>=0.33.0
transformers==4.50.0
diffusers==0.35.2
peft==0.17.1
accelerate==1.6.0

# ACE-Step Music Generation (from GitHub)
# Installed after core dependencies to use pinned versions above
git+https://github.com/ACE-Step/ACE-Step.git

# Audio Processing
demucs==4.0.1
pedalboard==0.9.19
pydub>=0.25.1
librosa==0.11.0
soundfile==0.13.1
audioread>=3.0.0

# UI - Match with sdk_version in README
gradio==5.49.1

# Dataset Management
huggingface_hub>=0.20.0
datasets>=2.16.0
requests>=2.31.0
tqdm>=4.66.0

# Utilities
numpy>=1.24.0
pyyaml>=6.0
loguru>=0.7.0
